loopdb: move code to shared code file
